Controllers Indexes Updated

+++ Core/CarbonEmissionCalculator.Domain/Entities/MobileOffRoadGasolineCalculation.cs
@@ -4,23 +4,23 @@ namespace CarbonEmissionCalculator.Domain.Entities
 {
     public class MobileOffRoadGasolineCalculation : BaseEntity
     {
-        public decimal ConsumptionAmount { get; set; }
-        public decimal Density { get; set; }
-        public decimal ConsumptionKg { get; set; }
-        public decimal ConsumptionGg { get; set; }
-        public decimal NetCaloricValue { get; set; }
-        public decimal ConsumptionTJ { get; set; }
-        public decimal CarbonOxidationFactor { get; set; }
-        public decimal Co2Factor { get; set; }
-        public decimal Ch4Factor { get; set; }
-        public decimal N2oFactor { get; set; }
-        public decimal Co2GWP { get; set; }
-        public decimal Ch4GWP { get; set; }
-        public decimal N2oGWP { get; set; }
-        public decimal TotalCO2 { get; set; }
-        public decimal TotalCH4 { get; set; }
-        public decimal TotalN2O { get; set; }
-        public decimal TotalCO2e { get; set; }
-        public decimal TotalCO2eTon { get; set; }
+        public double ConsumptionAmount { get; set; }
+        public double Density { get; set; }
+        public double ConsumptionKg { get; set; }
+        public double ConsumptionGg { get; set; }
+        public double NetCaloricValue { get; set; }
+        public double ConsumptionTJ { get; set; }
+        public double CarbonOxidationFactor { get; set; }
+        public double Co2Factor { get; set; }
+        public double Ch4Factor { get; set; }
+        public double N2oFactor { get; set; }
+        public double Co2GWP { get; set; }
+        public double Ch4GWP { get; set; }
+        public double N2oGWP { get; set; }
+        public double TotalCO2 { get; set; }
+        public double TotalCH4 { get; set; }
+        public double TotalN2O { get; set; }
+        public double TotalCO2e { get; set; }
+        public double TotalCO2eTon { get; set; }
     }
 }
